% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ncm.R
\name{print.summary_ncm}
\alias{print.summary_ncm}
\title{Print summary of NCM model}
\usage{
\method{print}{summary_ncm}(x, ...)
}
\arguments{
\item{x}{An object of class "summary_ncm".}

\item{...}{Additional arguments (not used).}
}
\description{
Print summary of NCM model
}
