% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ncm.R
\name{summary.NCM}
\alias{summary.NCM}
\title{Summary method for NCM objects}
\usage{
\method{summary}{NCM}(object, ...)
}
\arguments{
\item{object}{An object of class "NCM" (single model).}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class "summary_ncm" containing key model statistics.
}
\description{
Summary method for NCM objects
}
