% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ncm.R
\name{summary.NCM_group}
\alias{summary.NCM_group}
\title{Summary method for NCM group results}
\usage{
\method{summary}{NCM_group}(object, ...)
}
\arguments{
\item{object}{An object of class "NCM_group" (list of NCM objects).}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class "summary_ncm_group" containing summary for each group.
}
\description{
Summary method for NCM group results
}
