% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ncm.R
\name{print.summary_ncm_group}
\alias{print.summary_ncm_group}
\title{Print summary of NCM group results}
\usage{
\method{print}{summary_ncm_group}(x, ...)
}
\arguments{
\item{x}{An object of class "summary_ncm_group".}

\item{...}{Additional arguments (not used).}
}
\description{
Print summary of NCM group results
}
